Added abstract adapter for azure
